add stat and bug fix

+++ SQLWORK.py
@@ -1,6 +1,7 @@
 import sqlite3
 import os
-from datetime import date
+from datetime import date,timedelta,datetime
+import matplotlib.pyplot as plt
 
 class DataBase:
     def __init__(self):
@@ -96,7 +97,9 @@ class DataBase:
         try:
             ad = self.cur.execute(f""" SELECT * FROM AD_MAIN WHERE id='{id}';""").fetchone()
             photo = ad[4]
-            os.remove(os.getcwd() + photo)
+            if len(self.cur.execute(f""" SELECT * FROM AD WHERE photo='{photo}';""").fetchall()) == 0 and len(
+                    self.cur.execute(f""" SELECT * FROM AD_MAIN WHERE photo='{photo}';""").fetchall()) == 1:
+                os.remove(os.getcwd() + photo)
             self.cur.execute(f"""DELETE FROM AD_MAIN WHERE id='{id}';""")
             self.conn.commit()
         except:
@@ -106,7 +109,8 @@ class DataBase:
         try:
             ad = self.cur.execute(f""" SELECT * FROM AD WHERE id='{id}';""").fetchone()
             photo = ad[4]
-            os.remove(os.getcwd() + photo)
+            if len(self.cur.execute(f""" SELECT * FROM AD WHERE photo='{photo}';""").fetchall())==1 and len(self.cur.execute(f""" SELECT * FROM AD_MAIN WHERE photo='{photo}';""").fetchall())==0:
+                os.remove(os.getcwd() + photo)
             self.cur.execute(f"""DELETE FROM AD WHERE id='{id}';""")
             self.conn.commit()
             return [int(ad[7]), f'Ваше объявление {ad[2]}, {ad[5]}₽ не прошло модерацию!❌']
@@ -123,3 +127,40 @@ class DataBase:
             return [int(ad[7]),f'Ваше объявление {ad[2]}, {ad[5]}₽ прошло модерацию и опубликовано!✅']
         except:
             pass
+
+    async def get_stats(self):
+        users = self.cur.execute(f""" SELECT * FROM USERS;""").fetchall()
+        quantity_users = len(users)
+        mount_ago = date.today() - timedelta(days=31)
+        users_last_mounth = []
+        for i in users:
+            dat = i[1].split('-')
+            dat = date(int(dat[0]), int(dat[1]), int(dat[2]))
+            if dat >= mount_ago:
+                users_last_mounth.append(i)
+        users_last_mounth = len(users_last_mounth)
+        on_moder_ads = len(self.cur.execute(f""" SELECT * FROM AD;""").fetchall())
+        ads = len(self.cur.execute(f""" SELECT * FROM AD_MAIN;""").fetchall())
+
+        labels = ['Промышленное','Логистика и склад','Для магазина','Для ресторана','Для салона красоты','Лабораторное','Медицинское','Другое']
+        values = []
+        labelsn = []
+        for i in labels:
+            clv = len(self.cur.execute(f""" SELECT * FROM AD_MAIN WHERE type='{i}';""").fetchall())
+            if clv==0:
+                pass
+            else:
+                values.append(clv)
+                labelsn.append(i)
+        labels = labelsn
+        fig1, ax1 = plt.subplots()
+        wedges, texts, autotexts = ax1.pie(values, labels=labels, autopct='%1.2f%%')
+        ax1.axis('equal')
+        filename = str(datetime.utcnow()).replace(' ','_').replace('.',':',1).replace(':','-')
+        plt.savefig(os.getcwd()+f'\\images\\stats\\{filename}.png')
+        return [f'Пользователи: {quantity_users} ({users_last_mounth} за последние 31 день)\nОбъявлений: {ads}\nОбъявлений на модерации: {on_moder_ads}',filename]
+
+
+
+
+
